+++ src/components/org/jallen/common/Types.tsx
@@ -4,3 +4,8 @@ export interface Socials {
   name: string;
   href: string;
 }
+
+export interface Profiles {
+  id: number;
+  description: string;
+}