+++ src/components/org/jallen/header/Header.tsx
@@ -1,11 +1,12 @@
 import Link from 'next/link';
 import { Button } from '@/components/ui/button';
+import { FaArrowRight } from 'react-icons/fa';
 import Navigation from '../navigation/Navigation';
 import MobileNavigation from '../navigation/MobileNavigation';
 
 const Header = () => {
   return (
-    <header className='py-8 xl:py-12 text-white bg-pink-50/20'>
+    <header className='py-8 xl:py-12 text-white bg-gray-800'>
       <div className='container mx-auto flex justify-between items-center'>
         <Link href='/'>
           <h1 className='text-4xl font-semibold'>
@@ -13,15 +14,24 @@ const Header = () => {
           </h1>
         </Link>
 
-        <div className='hidden xl:flex items-center gap-8'>
+        <div className='hidden lg:flex items-center gap-8'>
           <Navigation />
-          <Link href='/contact'>
-            <Button>Hire me</Button>
+          <Link href='https://www.linkedin.com/in/tjamesallen15/'>
+            <Button className='gap-2 hover:bg-accent transition-all'>
+              Hire me
+              <FaArrowRight />
+            </Button>
           </Link>
         </div>
 
-        <div className='xl:hidden'>
+        <div className='lg:hidden flex flex-row items-center gap-4'>
           <MobileNavigation />
+          <Link href='https://www.linkedin.com/in/tjamesallen15/'>
+            <Button className='gap-2 hover:bg-accent transition-all'>
+              Hire me
+              <FaArrowRight />
+            </Button>
+          </Link>
         </div>
       </div>
     </header>