+++ src/components/org/jallen/implementation/ProjectsImpl.tsx
@@ -41,7 +41,7 @@ const ProjectsImpl = () => {
           animate={{ 
             opacity: 1,
             transition: { 
-              delay: 2.4,
+              delay: 1.5,
               duration: 0.4,
               ease: 'easeIn'
             }