+++ src/components/org/jallen/navigation/Navigation.tsx
@@ -24,7 +24,7 @@ const Navigation = () => {
     <nav className='flex gap-8'>
       { links.map((link, index) => {
         return (
-          <Link href={link.path} key={index} className={`${link.path === pathName && "text-accent border-b-2 border-accent"} capitalize font-medium hover:text-accent transition-all`}>
+          <Link href={link.path} key={index} className={`${link.path === pathName && "text-accent border-b-2 border-accent"} capitalize font-karla text-lg hover:text-accent transition-all`}>
             {link.name}
           </Link>
         )