+++ src/components/org/jallen/profile/Profile.tsx
@@ -11,7 +11,7 @@ const Profile = () => {
         animate={{ 
           opacity: 1,
           transition: { 
-            delay: 2,
+            delay: 0.5,
             duration: 0.4,
             ease: 'easeIn'
           }