+++ src/components/org/jallen/transition/StairTransition.tsx
@@ -18,7 +18,7 @@ const StairTransition = () => {
           initial={{ opacity: 1 }}
           animate={{ 
             opacity: 0,
-            transition: { delay: 1, duration: 0.4, ease: 'easeInOut' }
+            transition: { delay: 0.2, duration: 0.4, ease: 'easeInOut' }
           }}
         />
       </div>