+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+# 1.1.0
+
+## 1.1.0 Fixes
+
+`[Header]` Fixed design and styles.
+`[Footer]` Fixed design and styles.
+`[MobileNavigation]` Fixed design and styles.
+`[Navigation]` Fixed design and styles.
+`[Profile]` Fixed design and styles.
+`[Socials]` Fixed design and styles.
+`[Projects]` Fixed design and styles.
+`[Resume]` Fixed design and styles.
+
 # 1.0.0
 
 ## 1.0.0 Features